requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.11.11
aiosignal==1.3.1
aiosqlite==0.20.0
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
arrow==1.3.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==24.2.0
bcrypt==4.2.0
binaryornot==0.4.4
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
codecov==2.1.13
cookiecutter==2.6.0
coverage==7.6.10
cryptography==44.0.1
dependency-injector==4.42.0
Deprecated==1.2.15
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
fakeredis==2.24.1
fastapi==0.115.4
flet==0.27.6
flet-core==0.24.1
flet-desktop==0.27.6
flet-runtime==0.24.1
frozenlist==1.4.1
greenlet==3.0.3
h11==0.16.0
httpcore==1.0.9
httptools==0.6.1
httpx==0.27.2
idna==3.8
iniconfig==2.0.0
Jinja2==3.1.6
limits==4.0.0
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.0.5
mypy==1.16.1
mypy_extensions==1.1.0
oauthlib==3.2.2
packaging==23.2
passlib==1.7.4
pathspec==0.12.1
pluggy==1.5.0
propcache==0.2.1
psycopg2-binary==2.9.9
pyasn1==0.6.0
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.9.0
pypng==0.20220715.0
pytest==8.3.2
pytest-asyncio==0.24.0
pytest-cov==6.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.18
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
qrcode==7.4.2
redis==5.0.8
repath==0.9.0
requests==2.32.4
rich==13.8.0
rsa==4.9
ruff==0.12.2
six==1.16.0
slowapi==0.1.9
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.32
sse-starlette==2.1.3
starlette==0.40.0
text-unidecode==1.3
tomli==2.0.2
types-python-dateutil==2.9.0.20240821
typing_extensions==4.12.2
urllib3==2.5.0
uvicorn==0.30.6
uvloop==0.21.0
watchdog==4.0.2
watchfiles==0.24.0
websockets==13.0.1
wrapt==1.17.0
yarl==1.18.3
